+++ src/main/scala/com/lucendar/strm/common/strm/log/StrmMqLogEntry.java
@@ -1,0 +1,164 @@
+package com.lucendar.strm.common.strm.log;
+
+import java.util.StringJoiner;
+
+/**
+ * 消息队列日志项
+ */
+public class StrmMqLogEntry {
+
+    public static final String CHANNEL__RABBIT = "rabbit";
+    public static final String CHANNEL__APACHE_MQ = "apacheMq";
+    public static final String CHANNEL__ARTEMIS = "artemis";
+    public static final String CHANNEL__KAFKA = "kafka";
+    public static final String CHANNEL__MQTT = "mqtt";
+    public static final String CHANNEL__REDIS = "redis";
+
+    /**
+     * 事件类型：发送
+     */
+    public static final int TYP__SEND = 1;
+
+    /**
+     * 事件类型：接收
+     */
+    public static final int TYP__RECV = 2;
+
+    private String id;
+    private long tm;
+    private String appId;
+    private String channel;
+    private String destination;
+    private int typ;
+    private String msg;
+
+    /**
+     * 取日志记录ID
+     *
+     * @return 日志记录ID
+     */
+    public String getId() {
+        return id;
+    }
+
+    /**
+     * 设置日志记录ID
+     * @param id 日志记录ID
+     */
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    /**
+     * 取事件时间, epoch millis
+     * @return 事件时间, epoch millis
+     */
+    public long getTm() {
+        return tm;
+    }
+
+    /**
+     * 设置事件时间, epoch millis
+     * @param tm 事件时间, epoch millis
+     */
+    public void setTm(long tm) {
+        this.tm = tm;
+    }
+
+    /**
+     * 取 AppID
+     * @return AppID
+     */
+    public String getAppId() {
+        return appId;
+    }
+
+    /**
+     * 设置AppID
+     * @param appId AppID
+     */
+    public void setAppId(String appId) {
+        this.appId = appId;
+    }
+
+    /**
+     * 取消息队列类型。为 CHANNEL__xxx 系列常量之一
+     *
+     * @return 消息队列类型
+     */
+    public String getChannel() {
+        return channel;
+    }
+
+    /**
+     * 设置消息队列类型。为 CHANNEL__xxx 系列常量之一
+     * @param channel 消息队列类型。为 CHANNEL__xxx 系列常量之一
+     */
+    public void setChannel(String channel) {
+        this.channel = channel;
+    }
+
+    /**
+     * 取主题名称
+     *
+     * @return 主题名称
+     */
+    public String getDestination() {
+        return destination;
+    }
+
+    /**
+     * 设置主题名称
+     * @param destination 主题名称
+     */
+    public void setDestination(String destination) {
+        this.destination = destination;
+    }
+
+    /**
+     * 取事件类型，为 TYP__xxx 系列常量之一
+     *
+     * @return 事件类型，为 TYP__xxx 系列常量之一
+     */
+    public int getTyp() {
+        return typ;
+    }
+
+    /**
+     * 设置事件类型，为 TYP__xxx 系列常量之一
+     * @param typ 事件类型，为 TYP__xxx 系列常量之一
+     */
+    public void setTyp(int typ) {
+        this.typ = typ;
+    }
+
+    /**
+     * 取发送或接收的消息
+     *
+     * @return 发送或接收的消息
+     */
+    public String getMsg() {
+        return msg;
+    }
+
+    /**
+     * 设置发送或接收的消息
+     * @param msg 发送或接收的消息
+     */
+    public void setMsg(String msg) {
+        this.msg = msg;
+    }
+
+    @Override
+    public String toString() {
+        return new StringJoiner(", ", StrmMqLogEntry.class.getSimpleName() + "[", "]")
+                .add("id=" + id)
+                .add("tm=" + tm)
+                .add("appId='" + appId + "'")
+                .add("channel='" + channel + "'")
+                .add("destination='" + destination + "'")
+                .add("typ=" + typ)
+                .add("msg='" + msg + "'")
+                .toString();
+    }
+}